refactor(goose): improve envs key in goose config
The envs key in the goose config was previously defined as a map of
string to interface{}, but it should be a map of string to
string. This commit changes the type of the envs key to
map[string]string.  Additionally, the env_keys key was removed from
the fetch and pdf_reader extensions because it was not being used.