refactor(mitmproxy): move mitmproxy scripts and data to repo
This commit moves mitmproxy scripts and data files out of the
chezmoi managed home directory and into the repository, to avoid
inconsistencies.  This has the benefit of keeping the scripts
version controlled and ensures that the mitmproxy scripts are
available when running the mitmproxy container.

Additionally, the map-local.py script has been modified to load its
data files from the same folder where the script resides to provide
the correct path for the mitmproxy container.